Replace default.md with siteConvertLayout.njk
We replace default.md as we can possibly generate our desired default.md
from siteConvertLayout.njk. This can reduce the unexpected behaviours to
style changes and more as only one file will need to be updated instead
of two.

This should be adequete as the init template should be more focused
on a friendly onboarding of markbind's features rather than markbind's
default layout.